[Setting] #108 - 폴더 구조 수정
