+++ Diptych/Diptych/Data/Manager/FirebaseMananger.swift
