+++ Diptych/Diptych/Data/Manager/ImageCacheManager.swift
