+++ Diptych/Diptych/Presentation/Auth/Model/DiptychAlbum.swift
