+++ Diptych/Diptych/Resources/DiptychSplashLogo.json
