+++ Diptych/Diptych/Resources/DiptychSplashLogo3sec.json
